newsletter unsubscribe

+++ src/pages/api/newsletter/unsubscribe.ts
@@ -2,17 +2,38 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { prisma } from "../../../server/db";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const { id } = req.body;
+  const { query } = req;
+  const unSubEmail = Array.isArray(query.unSubEmail)
+    ? query.unSubEmail[0]
+    : query.unSubEmail;
+
+  if (!unSubEmail) {
+    return res.status(400).json({ message: "Missing email" });
+  }
+
+  console.log("sssssssss", unSubEmail);
 
   try {
-    const subscriber = await prisma.newsletterSubscriber.delete({
+    // check if email exists
+    const subscriberExists = await prisma.newsletterSubscriber.findFirst({
       where: {
-        id: id,
+        email: unSubEmail,
       },
     });
 
-    res.status(200).json(subscriber);
+    if (!subscriberExists) {
+      return res.status(400).json({ message: "Email does not exist" });
+    }
+    await prisma.newsletterSubscriber.delete({
+      where: {
+        email: unSubEmail,
+      },
+    });
+
+    res.status(200).json({ message: "Unsubscribed" });
   } catch (err) {
     console.log(err);
   }
 };
+
+export default handler;
